requirements.txt
-e .
accelerate
datasets
transformers
math-verify
vllm
wandb
hf_transfer
peft
#flash-attn --no-build-isolation
langchain_openai
python-dotenv